+++ lib/asn1constants.dart
@@ -7,6 +7,7 @@ const int INTEGER_TYPE = 0x02;
 const int BIT_STRING_TYPE = 0x03;
 const int OCTET_STRING_TYPE = 0x04;
 const int NULL_TYPE = 0x05;
+const int OBJECT_DESCRIPTOR_TYPE = 0x07;
 const int ENUMERATED_TYPE = 0x0A;
 const int UTF8_STRING_TYPE = 0x0C;
 const int NUMERIC_STRING_TYPE = 0x12;
@@ -14,8 +15,11 @@ const int PRINTABLE_STRING_TYPE = 0x13;
 const int IA5_STRING_TYPE = 0x16;
 const int UTC_TIME_TYPE = 0x17;
 const int BMP_STRING_TYPE = 0x1e;
-const int SEQUENCE_TYPE = 0x30;
-const int SET_TYPE = 0x31;
+// SET and SEQUENCE TYPES include the constructed bit set.
+const int SEQUENCE_TYPE = 0x10;
+const int SET_TYPE = 0x11;
+const int CONSTRUCTED_SEQUENCE_TYPE = CONSTRUCTED_BIT | SEQUENCE_TYPE;  // 0x10 + constructed bit
+const int CONSTRUCTED_SET_TYPE = CONSTRUCTED_BIT | SET_TYPE; // 0x11 + constructed bit
 const int OBJECT_IDENTIFIER = 0x06;
 const int GENERALIZED_TIME = 0x18;
 const int TELETEXT_STRING = 0x14;
@@ -26,3 +30,23 @@ const int BOOLEAN_FALSE_VALUE = 0x00;
 // application-specific tag bytes for SNMP protocol
 
 const int IP_ADDRESS = 0x40;
+
+// Generic application type.
+const int APPLICATION_TYPE = 0x40;
+// Private type.
+const int PRIVATE_TYPE = 0xc0;
+
+// Value to AND against to test for the constructed bit (sequence, set)
+const int CONSTRUCTED_BIT = 0x20;
+
+// Utilities to test the tag
+
+// return just the type bits - the lower 5 bits of the tag
+int type_bits(int tag) => tag & 0x1f;
+bool isPrimitive(int tag) => (0xC0 & tag) == 0;
+bool isApplication(int tag) => (APPLICATION_TYPE & tag) == APPLICATION_TYPE;
+bool isPrivate(int tag) => (PRIVATE_TYPE & tag)  == PRIVATE_TYPE;
+bool isConstructed(int tag) => CONSTRUCTED_BIT & tag != 0;
+// set or sequence
+bool isSet(int tag) => isConstructed(tag) && type_bits(tag) == SET_TYPE;
+bool isSequence(int tag) => isConstructed(tag) && type_bits(tag) == SEQUENCE_TYPE;